+++ banking/x64/Debug/banking.log
@@ -1,2 +1,4 @@
-﻿  main.cpp
+﻿  AccountHandler.cpp
+  BankingSystemMain.cpp
+  코드를 생성하고 있습니다...
   banking.vcxproj -> C:\Users\woodz\source\repos\banking\x64\Debug\banking.exe
